# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GST Invoice System API",
    "version": "1.0.0",
...
    "license": {
      "name": "See LICENSE in the repository"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    },
    {
      "url": "https://{host}",
      "description": "Your deployment",
      "variables": {
        "host": {
          "default": "invoices.example.com"
        }
      }
    }
  ],
...
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session",
        "description": "Set by POST /api/auth/login. HttpOnly, SameSite, and revocable server-side."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "Human-readable message, safe to show a user."
          },
          "code": {
            "type": "string",
            "description": "Stable machine-readable code, present on errors a client is expected to handle (PLAN_LIMIT, PERIOD_LOCKED, VALUATION_RESTATEMENT, RETENTION_WARNING, DUPLICATE, HAS_MOVEMENTS, ALREADY_POSTED)."
          }
        }
      },
      "ValidationError": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Error"
          },
          {
            "type": "object",
            "properties": {
              "issues": {
                "type": "array",
                "description": "Zod issues, when the failure was schema validation.",
                "items": {
                  "type": "object",
                  "properties": {
                    "path": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        ]
      },
      "Paise": {
        "type": "integer",
        "format": "int64",
        "description": "An amount in integer PAISE (1/100 rupee). Every monetary field in this API is paise and is named with a `Paise` suffix. Rupees are never sent over the wire: they cannot represent 1/3 of a rupee exactly, and accumulating them loses money.",
        "example": 118000
      },
      "Quantity": {
        "type": "number",
        "description": "A stock quantity. Fractional, because real units are (2.5 kg, 0.75 m). Not money - quantities are the only numbers in this API that are not integers.",
        "example": 2.5
      },
...
      "TaxTotals": {
        "type": "object",
        "description": "Computed server-side. Values sent by a client are ignored.",
        "properties": {
          "subTotalPaise": {
            "$ref": "#/components/schemas/Paise"
          },
          "discountPaise": {
            "$ref": "#/components/schemas/Paise"
          },
          "cgstTotalPaise": {
            "$ref": "#/components/schemas/Paise"
          },
          "sgstTotalPaise": {
            "$ref": "#/components/schemas/Paise"
          },
          "igstTotalPaise": {
            "$ref": "#/components/schemas/Paise"
          },
          "cessTotalPaise": {
            "$ref": "#/components/schemas/Paise"
          },
          "taxTotalPaise": {
            "$ref": "#/components/schemas/Paise"
          },
          "tdsPaise": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Paise"
              }
            ],
            "description": "Shown separately and NOT deducted from grandTotalPaise. TDS is withheld by the buyer from payment; the invoice is still for the full amount."
          },
          "roundOffPaise": {
            "$ref": "#/components/schemas/Paise"
          },
          "grandTotalPaise": {
            "$ref": "#/components/schemas/Paise"
          },
          "isInterState": {
            "type": "boolean",
            "description": "True means IGST; false means CGST + SGST. Never both."
          },
          "reverseCharge": {
            "type": "boolean"
          },
          "placeOfSupply": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "Invoice": {
        "allOf": [
          {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "number": {
                "type": "string",
                "example": "INV/26-27/0042",
                "description": "Allocated atomically inside the same transaction as the document, per financial year. GST requires a gap-free consecutive series."
              },
              "date": {
                "type": "string",
                "format": "date-time"
              },
              "dueDate": {
                "type": "string",
                "format": "date-time",
                "nullable": true
              },
              "status": {
                "type": "string",
                "enum": [
                  "UNPAID",
                  "PARTIAL",
                  "PAID",
                  "CANCELLED"
                ]
              },
              "partyId": {
                "type": "string"
              },
              "amountPaidPaise": {
                "$ref": "#/components/schemas/Paise"
              },
              "notes": {
                "type": "string",
                "nullable": true
              },
              "items": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/InvoiceLine"
                }
              }
            }
          },
          {
            "$ref": "#/components/schemas/TaxTotals"
          }
        ]
      },
      "InvoiceLine": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "itemId": {
            "type": "string",
            "nullable": true
          },
          "itemName": {
            "type": "string"
          },
          "hsn": {
            "type": "string",
            "nullable": true
          },
          "quantity": {
            "$ref": "#/components/schemas/Quantity"
          },
          "unit": {
            "type": "string"
          },
          "ratePaise": {
            "$ref": "#/components/schemas/Paise"
          },
          "discountPaise": {
            "$ref": "#/components/schemas/Paise"
          },
          "apportionedDiscountPaise": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Paise"
              }
            ],
            "description": "Share of an invoice-level discount allocated to this line by taxable value, so the line taxable values still sum to the invoice taxable value exactly."
          },
          "taxablePaise": {
            "$ref": "#/components/schemas/Paise"
          },
          "gstRate": {
            "type": "number"
          },
          "cgstPaise": {
            "$ref": "#/components/schemas/Paise"
          },
          "sgstPaise": {
            "$ref": "#/components/schemas/Paise"
          },
          "igstPaise": {
            "$ref": "#/components/schemas/Paise"
          },
          "cessPaise": {
            "$ref": "#/components/schemas/Paise"
          },
          "totalPaise": {
            "$ref": "#/components/schemas/Paise"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "sessionCookie": []
    }
  ],
  "paths": {
...
    "/api/quotations/{id}/convert": {
      "post": {
        "tags": [
          "Sales"
        ],
        "summary": "Convert a quotation to an invoice",
        "description": "Moves stock and posts the ledger at this point, not when the quotation was raised.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Resource id. Always scoped to the caller's company."
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Invoice"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in, or the session was revoked",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Read-only role, or the company is suspended",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found, or belongs to another company",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/stock-transfers": {
      "get": {
        "tags": [
          "Inventory"
        ],
        "summary": "List stock transfers",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not signed in, or the session was revoked",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Inventory"
        ],
        "summary": "Transfer stock between godowns",
        "description": "Writes a matched out/in pair at the SAME cost: valuing the receipt independently would create or destroy value by moving goods between your own shelves.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "itemId",
                  "fromGodownId",
                  "toGodownId",
                  "quantity"
                ],
                "properties": {
                  "itemId": {
                    "type": "string"
                  },
                  "fromGodownId": {
                    "type": "string"
                  },
                  "toGodownId": {
                    "type": "string"
                  },
                  "quantity": {
                    "type": "number"
                  },
                  "notes": {
                    "type": "string"
                  },
                  "date": {
                    "type": "string",
                    "format": "date-time"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in, or the session was revoked",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "402": {
            "description": "Requires the Basic plan",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Read-only role, or the company is suspended",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}